requirements.txt
dj-database-url==0.5.0
Django==2.2.1
django-bootstrap3
django-environ
django-filter
django-rainbowtests==0.6.0
django-heroku
django-tables2
gunicorn==19.9.0
psycopg2-binary==2.7.7
Pygments==2.4.0
pytz==2019.1
sqlparse==0.3.0
whitenoise==4.1.2
